+++ default.py
@@ -1,21 +1,24 @@
 import xbmcgui
+import xbmc
 
 #Try again later
-# dummy_list = [
-
-#         {'MediaType':'Movie', 'Name': 'Movie 1', 'Rating': '5', 'Genre': 'Horror', 'Length': '130', 'Year': '2019'},
-#         {'MediaType':'Epsiode', 'Name': 'Epsiode 2', 'Rating': '2', 'Genre': 'Slice-Of-Life', 'Length': '40', 'Year': '2020'},
-#         {'MediaType':'Movie', 'Name': 'Movie 2', 'Rating': '4', 'Genre': 'Comedy', 'Length': '90', 'Year': '2023'},
-#         {'MediaType':'Movie', 'Name': 'Movie 3', 'Rating': '3', 'Genre': 'Action', 'Length': '120', 'Year': '2022'},
-#         {'MediaType':'Episode', 'Name': 'Epsode 2', 'Rating': '1', 'Genre': 'Romance', 'Length': '20', 'Year': '2021'}
-
-#         ]
-
+dummy_list = [
+        {'MediaType':'Movie', 'Name': 'Movie 1', 'Rating': '5', 'Genre': 'Horror', 'Length': '130', 'Year': '2019'},
+        {'MediaType':'Epsiode', 'Name': 'Epsiode 2', 'Rating': '2', 'Genre': 'Slice-Of-Life', 'Length': '40', 'Year': '2020'},
+        {'MediaType':'Movie', 'Name': 'Movie 2', 'Rating': '4', 'Genre': 'Comedy', 'Length': '90', 'Year': '2023'},
+        {'MediaType':'Movie', 'Name': 'Movie 3', 'Rating': '3', 'Genre': 'Action', 'Length': '120', 'Year': '2022'},
+        {'MediaType':'Episode', 'Name': 'Epsode 2', 'Rating': '1', 'Genre': 'Romance', 'Length': '20', 'Year': '2021'}
+]
 
 class MyWindow(xbmcgui.Window):
     def __init__(self) -> None:
         super().__init__()
+        self.button = xbmcgui.ControlButton(250, 250, 250, 250, "Display List")
+        self.addControl(self.button)
 
+    def display_list(self):
+            dialog = xbmcgui.Dialog()
+            dialog.ok("List", str(dummy_list))
 
      #The Identifying stuff
     def media_Type(self, text):
@@ -50,13 +53,54 @@ class MyWindow(xbmcgui.Window):
         label = xbmcgui.ControlLabel(1120, 50, 200, 100, text)
         self.addControl(label)
 
+    def mostWatched(self, text):
+        label = xbmcgui.ControlLabel( 37, 200, 200, 100, text)
+        self.addControl(label)
+
+    def casts(self, text):
+        label = xbmcgui.ControlLabel(225, 200, 200, 100, text)
+        self.addControl(label)
+
+    def director(self, text):
+        label = xbmcgui.ControlLabel(360, 200, 300, 200, text)
+        self.addControl(label)
+
+
+    def playOne(self, text):
+        label = xbmcgui.ControlLabel(1130, 660, 300, 200, text)
+        self.addControl(label)
+
+    def showList(self, text):
+        label = xbmcgui.ControlLabel(935, 660, 300, 200, text)
+        self.addControl(label)
+
+    def loopPlay(self, text):
+        label = xbmcgui.ControlLabel(725, 660, 300, 200, text)
+        self.addControl(label)
+
+    def saveSearch(self, text):
+        label = xbmcgui.ControlLabel(555, 660, 300, 200, text)
+        self.addControl(label)
+
+    def shutDown(self, text):
+        label = xbmcgui.ControlLabel(95, 610, 300, 200, text)
+        self.addControl(label)
+
+    def time(self, text):
+        label = xbmcgui.ControlLabel(95, 660, 300, 200, text)
+        self.addControl(label)
+    
 
        #The Orange
     def show_circle(self, x, y, radius, color):
         circle = xbmcgui.ControlImage(x, y, radius, radius, "/Users/collegecarol/Desktop/Kodi/Kodi4/circle.png", color)
         self.addControl(circle)
 
-    
+    def onClick(self, controlId: int):
+        xbmc.log(f"something was clicked! {controlId}")
+        if controlId == self.button.getId():
+            xbmc.log("the DisplayList button was pressed!")
+            self.display_list()
 
     # def get_names(items):
     # list = []
@@ -76,11 +120,9 @@ class MyWindow(xbmcgui.Window):
     # for param in dummy_list:
     #     first_params.append(param[0])
 
- 
 
   #The Window
 if __name__ == "__main__":
-    
     window = MyWindow()
 
     #Orange
@@ -115,13 +157,21 @@ if __name__ == "__main__":
     window.year("Year")
     window.tags("Tags")
     window.studio("Studio")
+    window.mostWatched("MostWatched")
+    window.casts("Casts")
+    window.director("Director")
 
-    #Try again Later
-    # button = window.getControl(200)
-    # button.setLabel("Display List")
-    # def display_list():
-    #     dialog = xbmcgui.Dialog()
-    #     dialog.ok("List", str(dummy_list))
+    window.playOne("Play One")
+    window.loopPlay("Loop Play")
+    window.showList("Show List")
+    window.saveSearch("Save")
+    window.shutDown("ShutDown Time")
+    window.time("Hour          Min")
+    
+
+
+    # Try again Later
+    # chatgpt being hallucinating again
     # button.setNavigation( xbmcgui.ACTION_SELECT_ITEM, lambda: display_list() )
 
     window.doModal()